extract env
